Refactoring backend components into packages.
