+++ app/controllers/GraphQLController.scala
@@ -1,8 +1,8 @@
 package controllers
 
+import components.Backend
 import javax.inject._
 import models.entities.Violations._
-import models.Backend
 import models.gql.GQLSchema
 import play.api.libs.json._
 import play.api.mvc._
