+++ app/controllers/HomeController.scala
@@ -1,7 +1,7 @@
 package controllers
 
+import components.Backend
 import javax.inject._
-import models.Backend
 import play.api.mvc._
 
 import scala.concurrent.ExecutionContext
@@ -14,7 +14,7 @@ class HomeController @Inject()(implicit
                               ) extends AbstractController(cc) {
 
   // example from here https://github.com/nemoo/play-slick3-example/blob/master/app/controllers/Application.scala
-  def index() =
+  def index(): Action[AnyContent] =
     Action { _ =>
       Ok(views.html.index())
     }
