+++ conf/application.conf
@@ -3,7 +3,7 @@ play.http.secret.key = "changeme"
 
 slick.dbs {
   default {
-    profile = "clickhouse.ClickHouseProfile$"
+    profile = "components.clickhouse.ClickHouseProfile$"
     db {
       driver = "ru.yandex.clickhouse.ClickHouseDriver"
       url = "jdbc:clickhouse://127.0.0.1:8123/ot"
@@ -12,7 +12,7 @@ slick.dbs {
     }
   }
   sumstats {
-    profile = "clickhouse.ClickHouseProfile$"
+    profile = "components.clickhouse.ClickHouseProfile$"
     db {
       driver = "ru.yandex.clickhouse.ClickHouseDriver"
       url = "jdbc:clickhouse://127.0.0.1:8123/ot"
