+++ production.conf
@@ -14,7 +14,7 @@ logger.application=INFO
 
 slick.dbs {
   default {
-    profile = "clickhouse.ClickHouseProfile$"
+    profile = "components.clickhouse.ClickHouseProfile$"
     db {
       driver = "ru.yandex.clickhouse.ClickHouseDriver"
       url = "jdbc:clickhouse://clickhouse-200201.c.open-targets-genetics.internal:8123/ot"
@@ -24,7 +24,7 @@ slick.dbs {
     }
   }
   sumstats {
-    profile = "clickhouse.ClickHouseProfile$"
+    profile = "components.clickhouse.ClickHouseProfile$"
     db {
       driver = "ru.yandex.clickhouse.ClickHouseDriver"
       url = "jdbc:clickhouse://clickhouse-200201.c.open-targets-genetics.internal:8123/ot"
@@ -45,5 +45,5 @@ ot.elasticsearch {
 # http.port=${?PLAY_PORT}
 # play.http.secret.key=${?PLAY_SECRET}
 # slick.dbs.default.db.url=${?SLICK_CLICKHOUSE_URL}
-# ot.elasticsearch.host=${?ELASTICSEARCH_HOST}
+# ot.components.elasticsearch.host=${?ELASTICSEARCH_HOST}
 # slick.dbs.default.db.url=${?SLICK_CLICKHOUSE_URL_SS}
