+++ test/models/SeqRepSpec.scala
@@ -1,17 +1,21 @@
 package models
 
-import clickhouse.rep.SeqRep._
+import components.clickhouse.rep.SeqRep._
 import org.scalatestplus.play._
 import org.scalatestplus.play.guice._
 import play.api.test._
 
 /**
-  * Add your spec here.
-  * You can mock out a whole application including requests, plugins etc.
-  *
-  * For more information, see https://www.playframework.com/documentation/latest/ScalaTestingWithScalaTest
-  */
-class FunctionsSpec extends PlaySpec with GuiceOneAppPerTest with Injecting {
+ * Add your spec here.
+ * You can mock out a whole application including requests, plugins etc.
+ *
+ * For more information, see https://www.playframework.com/documentation/latest/ScalaTestingWithScalaTest
+ */
+class SeqRepSpec extends PlaySpec with GuiceOneAppPerTest with Injecting {
+
+  "A Clickhouse Array(Int32) must be non-empty when 1 element present" in {
+    ISeqRep("[1]").rep mustBe Seq(1)
+  }
 
   "A Clickhouse Array(String) must be Seq.empty" in {
     StrSeqRep("['']").rep mustBe Seq.empty
